requirements.txt
# should change
cycler==0.10.0
future==0.18.2
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.0
opencv-python==4.2.0.34
openslide-python==1.1.1
Pillow==7.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.7.0
six==1.16.0
threadpoolctl==2.1.0
tifffile==2021.6.14
torch==1.9.1
torchvision==0.6.1
